Assignments/A05/requirements.txt
pymongo
uvicorn
fastapi
rich
python-dotenv
requests
pydantic